requirements.txt
python-telegram-bot 
pymupdf 
supabase 
requests 
python-dotenv
numpy
pytesseract 
pillow
langchain